outsourced routing in order to keep app.tsx clean; not lazy yet!
