+++ libs/router/README.md
@@ -1,0 +1,7 @@
+# router
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test router` to execute the unit tests via [Jest](https://jestjs.io).
