+++ libs/router/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  displayName: 'router',
+  preset: '../../jest.preset.js',
+  transform: {
+    '^.+\\.[tj]sx?$': 'babel-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../coverage/libs/router',
+};
