+++ libs/router/src/lib/router.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import Router from './router';
+
+describe('Router', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Router />);
+    expect(baseElement).toBeTruthy();
+  });
+});
