+++ tsconfig.base.json
@@ -23,6 +23,7 @@
       "@alpsbte/faq": ["libs/faq/src/index.ts"],
       "@alpsbte/gallery": ["libs/gallery/src/index.ts"],
       "@alpsbte/home": ["libs/home/src/index.ts"],
+      "@alpsbte/router": ["libs/router/src/index.ts"],
       "@alpsbte/shared/components": ["libs/shared/components/src/index.ts"]
     }
   },
